优化interface
