+++ packages/chameleon-tool/configs/mvvm/getExtendConfig.js
@@ -4,6 +4,9 @@ const getCommonConfig = require('../getCommonConfig');
 const utils = require('../utils.js');
 const {MvvmGraphPlugin} = require('mvvm-pack');
 const resolve = require('resolve');
+const originSourceLoader = {
+  loader: path.join(__dirname, './originSourceLoader.js')
+};
 
 module.exports = function(options) {
   let {type, media} = options;
@@ -15,6 +18,15 @@ module.exports = function(options) {
     media
   });
   cml.extPlatformPlugin[type] = platformPlugin;
+
+  function getCmlLoaders() {
+    let loaders = utils.cssLoaders({type, media});
+    loaders.js = [
+      loaders.js,
+      originSourceLoader
+    ]
+    return loaders;
+  }
   let extendConfig = {
     entry: {
       app: path.join(cml.projectRoot, 'src/app/app.cml')
@@ -30,7 +42,7 @@ module.exports = function(options) {
           use: [{
             loader: 'mvvm-cml-loader',
             options: {
-              loaders: utils.cssLoaders({type, media}),
+              loaders: getCmlLoaders(),
               cmlType: type,
               media,
               check: cml.config.get().check
@@ -46,12 +58,24 @@ module.exports = function(options) {
       }, platformPlugin)
     ]
   };
-  // options.moduleIdType = 'hash';
   let commonConfig = getCommonConfig(options);
   commonConfig.module.rules.forEach(item => {
+    // 静态资源的处理
     if (~['chameleon-url-loader', 'file-loader'].indexOf(item.loader)) {
       item.loader = 'mvvm-file-loader';
       item.options.publicPath = commonConfig.output.publicPath
+    }
+
+    if (item.test instanceof RegExp) {
+      // interface获取originSource
+      if (item.test.test('name.interface')) {
+        item.use.splice(1, 0, originSourceLoader)
+      }
+
+      // js获取originSource
+      if (item.test.test('name.js')) {
+        item.use.push(originSourceLoader)
+      }
     }
   })
