--- packages/chameleon-tool/configs/mvvm/utils.js
@@ -1,1 +1,0 @@
-
